+++ wp_api/src/api_client.rs
@@ -11,6 +11,9 @@ use crate::{
             application_passwords_endpoint::{
                 ApplicationPasswordsRequestBuilder, ApplicationPasswordsRequestExecutor,
             },
+            block_directory_endpoint::{
+                BlockDirectoryRequestBuilder, BlockDirectoryRequestExecutor,
+            },
             block_patterns_endpoint::{BlockPatternsRequestBuilder, BlockPatternsRequestExecutor},
             block_types_endpoint::{BlockTypesRequestBuilder, BlockTypesRequestExecutor},
             comments_endpoint::{CommentsRequestBuilder, CommentsRequestExecutor},
@@ -68,6 +71,7 @@ use std::sync::Arc;
 pub struct WpApiRequestBuilder {
     api_root: Arc<ApiRootRequestBuilder>,
     application_passwords: Arc<ApplicationPasswordsRequestBuilder>,
+    block_directory: Arc<BlockDirectoryRequestBuilder>,
     block_patterns: Arc<BlockPatternsRequestBuilder>,
     block_types: Arc<BlockTypesRequestBuilder>,
     autosaves: Arc<AutosavesRequestBuilder>,
@@ -113,6 +117,7 @@ impl WpApiRequestBuilder {
             auth_provider;
             api_root,
             application_passwords,
+            block_directory,
             block_patterns,
             block_types,
             autosaves,
@@ -168,6 +173,7 @@ impl UniffiWpApiClient {
 pub struct WpApiClient {
     api_root: Arc<ApiRootRequestExecutor>,
     application_passwords: Arc<ApplicationPasswordsRequestExecutor>,
+    block_directory: Arc<BlockDirectoryRequestExecutor>,
     block_patterns: Arc<BlockPatternsRequestExecutor>,
     block_types: Arc<BlockTypesRequestExecutor>,
     autosaves: Arc<AutosavesRequestExecutor>,
@@ -210,6 +216,7 @@ impl WpApiClient {
             delegate;
             api_root,
             application_passwords,
+            block_directory,
             block_patterns,
             block_types,
             autosaves,
@@ -262,6 +269,7 @@ pub trait IsWpApiClientDelegate {
 
 api_client_generate_endpoint_impl!(WpApi, api_root);
 api_client_generate_endpoint_impl!(WpApi, application_passwords);
+api_client_generate_endpoint_impl!(WpApi, block_directory);
 api_client_generate_endpoint_impl!(WpApi, block_patterns);
 api_client_generate_endpoint_impl!(WpApi, block_types);
 api_client_generate_endpoint_impl!(WpApi, autosaves);