+++ wp_api/src/api_error.rs
@@ -203,6 +203,8 @@ pub enum WpErrorCode {
     AlreadyTrashed,
     #[serde(rename = "rest_application_password_not_found")]
     ApplicationPasswordNotFound,
+    #[serde(rename = "rest_block_directory_cannot_view")]
+    BlockDirectoryCannotView,
     #[serde(rename = "rest_block_type_cannot_view")]
     BlockTypeCannotView,
     #[serde(rename = "rest_block_type_invalid")]